requirements.txt
edge-tts==6.1.7
fairseq==0.12.2
faiss-cpu==1.7.4
gradio==3.38.0
librosa==0.9.1
numpy==1.23.5
praat-parselmouth==0.4.3
ollama==0.3.3
PyAudio==0.2.14
PyGetWindow==0.0.9
PySide6==6.8.0
pyworld==0.3.4
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.4
torchcrepe==0.0.20
Unidecode==1.3.8
win11toast==0.35
winapps==0.2.0